Split snapshot_utils.rs into snapshot_bank_utils.rs
