+++ runtime/src/snapshot_config.rs
@@ -1,5 +1,8 @@
 use {
-    crate::snapshot_utils::{self, ArchiveFormat, SnapshotVersion},
+    crate::{
+        snapshot_bank_utils,
+        snapshot_utils::{self, ArchiveFormat, SnapshotVersion},
+    },
     solana_sdk::clock::Slot,
     std::{num::NonZeroUsize, path::PathBuf},
 };
@@ -50,9 +53,9 @@ impl Default for SnapshotConfig {
         Self {
             usage: SnapshotUsage::LoadAndGenerate,
             full_snapshot_archive_interval_slots:
-                snapshot_utils::DEFAULT_FULL_SNAPSHOT_ARCHIVE_INTERVAL_SLOTS,
+                snapshot_bank_utils::DEFAULT_FULL_SNAPSHOT_ARCHIVE_INTERVAL_SLOTS,
             incremental_snapshot_archive_interval_slots:
-                snapshot_utils::DEFAULT_INCREMENTAL_SNAPSHOT_ARCHIVE_INTERVAL_SLOTS,
+                snapshot_bank_utils::DEFAULT_INCREMENTAL_SNAPSHOT_ARCHIVE_INTERVAL_SLOTS,
             full_snapshot_archives_dir: PathBuf::default(),
             incremental_snapshot_archives_dir: PathBuf::default(),
             bank_snapshots_dir: PathBuf::default(),
